Move context menu code to PageBase
