+++ apps/opencs/view/prefs/keybindingpage.cpp
@@ -9,7 +9,6 @@
 
 #include "../../model/prefs/setting.hpp"
 #include "../../model/prefs/category.hpp"
-#include "../../view/prefs/contextmenuwidget.hpp"
 
 namespace CSVPrefs
 {
@@ -20,7 +19,7 @@ namespace CSVPrefs
         , mPageSelector(0)
     {
         // Need one widget for scroll area
-        CSVPrefs::ContextMenuWidget* topWidget = new CSVPrefs::ContextMenuWidget(category.getKey());
+        QWidget* topWidget = new QWidget();
         QVBoxLayout* topLayout = new QVBoxLayout(topWidget);
 
         // Allows switching between "pages"
